pyproject.toml
[project]
name = "cv-api"
version = "0.1.0"
description = "Computer vision API with FastAPI and ResNet18 on CIFAR-10"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.135.1",
    "uvicorn>=0.41.0",
    "pydantic>=2.12.0",
    "pydantic-settings>=2.3.0",
    "python-dotenv>=1.0.0",
    "torch",
    "torchvision",
    "pillow>=10.0.0",
]

